Fix PillSheetModifiedHistoryDeletedPillSheetAction
